Download modal mostly refactored
