+++ src/components/modals/modalHeader.js
@@ -1,0 +1,41 @@
+import {withStyles} from "@material-ui/core/styles";
+import MuiDialogTitle from "@material-ui/core/DialogTitle";
+import Typography from "@material-ui/core/Typography";
+import IconButton from "@material-ui/core/IconButton";
+import CloseIcon from "@material-ui/icons/Close";
+import React from "react";
+
+const styles = theme => ({
+  dialogTitle: {
+    margin: 0,
+    backgroundColor: 'rgba(229, 123, 0,1)',
+    padding: 20,
+    display: 'flex',
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    flexBasis: 50,
+    color: "white"
+  },
+  closeButton: {
+    color: "white",
+  }
+});
+
+const DialogTitle = withStyles(styles)(props => {
+  const { children, classes, onClose, ...other } = props;
+  return (
+    <MuiDialogTitle disableTypography className={classes.dialogTitle} {...other}>
+      <Typography variant="h6">
+        {children}
+      </Typography>
+      {onClose ? (
+        <IconButton aria-label="close" className={classes.closeButton} onClick={onClose}>
+          <CloseIcon />
+        </IconButton>
+      ) : null}
+    </MuiDialogTitle>
+  );
+});
+
+export default DialogTitle;
