+++ src/components/widgets/DownloadActionButton.js
@@ -1,5 +1,5 @@
 /** @jsx jsx */
-import {Fragment} from "react";
+import {Fragment, useState} from "react";
 import { useSelector, useDispatch } from "react-redux";
 import { jsx } from "@emotion/core";
 import textStyles from "../../styles/text";
@@ -24,6 +24,8 @@ export default function DownloadActionButton() {
     state.authorization.apiTokens[state.targets.source.name]);
   const selectedInSource = useSelector(state => state.resources.selectedInSource);
 
+  const [modalState, setModalState] = useState(false);
+
   /**
    * Open the modal.
    * Dispatch the downloadResource action
@@ -31,17 +33,15 @@ export default function DownloadActionButton() {
    *     -> On complete saga dispatches the ResourceDownloadSuccess action
    **/
   const submitDownload = () => {
-    toggleModalVisibility();
+    setModalState(true);
     dispatch(actionCreators.resources.downloadResource(selectedInSource, sourceTargetToken));
   };
-
-  const { modalVisible, toggleModalVisibility } = useModal();
 
   return (
     <Fragment>
       <DownloadModal
-        modalActive={modalVisible}
-        toggleModal={toggleModalVisibility}
+        modalState={modalState}
+        setModalState={setModalState}
       />
 
       <ActionButton
